+++ __tests__/controllers/engines/engine.router.test.js
@@ -1,8 +1,7 @@
 // module
-const { _fn } = require('../../../controllers/module/helpers');
+const { _$ } = require('../../../controllers/modules/');
 const { asyncWrapper, getMethods,
-    getMiddlewareQueue, getControllerQueue, getViewRenderQueue } = require('../../../controllers/engine/router')._test;
-
+    getMiddlewareQueue, getControllerQueue, getViewRenderQueue } = require('../../../controllers/engines/router')._test;
 
 
 // test
@@ -18,8 +17,8 @@ describe('Fn: asyncWrapper', () => {
         const mockInput = [async (req, res, next) => {}, (req, res, next) => {}];
         const result = mockInput.map(fn => asyncWrapper(fn));
         //
-        expect(_fn.object.checkNativeBrand(result[0], 'array')).toBeTruthy();
-        expect(_fn.object.checkNativeBrand(result[1], 'array')).toBeTruthy();
+        expect(_$.object.checkNativeBrand(result[0], 'array')).toBeTruthy();
+        expect(_$.object.checkNativeBrand(result[1], 'array')).toBeTruthy();
         expect(result[0].toString()).toBe('(req, res, next) => fn(req, res, next).catch(next)');
         expect(result[1].toString()).toBe('(req, res, next) => {}');
     });