+++ controllers/middleware/initial.js
@@ -1,13 +1,12 @@
-const { _fn } = require('../module/helpers');
+const { _$ } = require('../modules/');
 const { settingModel, postModel } = require('../../models/');
-const { ServerError } = require('../module/errors');
 
 
 
 const generic = async (req, res, next) => {
     // website settings
     const config = await settingModel.findOne({ active: true });
-    if (!config) throw new ServerError('No website configs, please restart the server for DB initialization.');
+    if (!config) throw new _$.error.ServerError('No website configs, please restart the server for DB initialization.');
     res.locals._site = config._doc;
 
     // connecting session
@@ -29,17 +28,17 @@ const postNormalizer = async (req, res, next) => {
     const post = req.body.post;
     if (!post) return next();
 
-    if (req.method === 'POST' && !post.canonical) post.canonical = _fn.string.toKebabCase(post.title);
+    if (req.method === 'POST' && !post.canonical) post.canonical = _$.string.toKebabCase(post.title);
     if (post.canonical !== undefined) {
         const canonicalCounts = await postModel.count({ canonical: post.canonical });
         if (canonicalCounts > 0) post.canonical = post.canonical + '-' + (canonicalCounts + 1);
     }
-    post.featured   = _fn.string.inspectFileURL(post.featured, res.locals._site.sets.imageTypes, { raw: false });
-    post.title      = _fn.string.escapeChars(post.title);
-    post.content    = _fn.string.escapeChars(post.content);                             // tofix: using sanitizer
-    post.category   = _fn.string.toKebabCase(post.category) || undefined;
-    post.tag        = _fn.string.toKebabCase(post.tag) || undefined;
-    post.visibility = _fn.object.assignDeep({}, post.visibility || 'normal', true);     // todo: split the assignment
+    post.featured   = _$.string.inspectFileURL(post.featured, res.locals._site.sets.imageTypes, { raw: false });
+    post.title      = _$.string.escapeChars(post.title);
+    post.content    = _$.string.escapeChars(post.content);                             // tofix: using sanitizer
+    post.category   = _$.string.toKebabCase(post.category) || undefined;
+    post.tag        = _$.string.toKebabCase(post.tag) || undefined;
+    post.visibility = _$.object.assignDeep({}, post.visibility || 'normal', true);     // todo: split the assignment
     return next();
 };