+++ controllers/routers/admin.js
@@ -6,7 +6,7 @@ module.exports = home = {};
 //  DEPENDENCIES
 // ==============================
 const { _md } = require('../middleware/plugins');
-const { settingModel, mediaModel, userModel } = require('../../models/index');
+const { settingModel, mediaModel, userModel } = require('../../models/');