+++ controllers/routers/post.js
@@ -5,8 +5,8 @@ module.exports = exports = { editor: {}, post: {} };
 // ==============================
 //  DEPENDENCIES
 // ==============================
-const { _fn } = require('../module/helpers');
-const { postModel } = require('../../models/index');
+const { _$ } = require('../modules/');
+const { postModel } = require('../../models/');
 const search = require('../middleware/search');
 
 
@@ -29,16 +29,16 @@ exports.editor.edit = {
         return res.redirect(`/post/editor/${req.session.view.post._id}`);
     },
     get: async (req, res, next) => {
-        if (!req.session.view) req.session.view = { post: await postModel.findById(_fn.string.readMongoId(req.url)) };
+        if (!req.session.view) req.session.view = { post: await postModel.findById(_$.string.readMongoId(req.url)) };
         return next();
     },
     patch: async (req, res) => {
-        const doc = await postModel.findByIdAndUpdate(_fn.string.readMongoId(req.url), req.body.post, { new: true });
+        const doc = await postModel.findByIdAndUpdate(_$.string.readMongoId(req.url), req.body.post, { new: true });
         req.flash('info', 'post have been successfully updated!');
         return res.redirect(`/post/${doc.canonical}`);
     },
     delete: async (req, res) => { // todo: trash can || double check
-        await postModel.postsDeleteThenDissociate(_fn.string.readMongoId(req.url), req.user);
+        await postModel.postsDeleteThenDissociate(_$.string.readMongoId(req.url), req.user);
         req.flash('info', 'post have been successfully deleted!');
         return res.redirect('/post/');
     },